refactor db queries
