+++ api/src/comments/routes.rs
@@ -21,7 +21,7 @@ use db::{
     comment::{self, Comment},
     user_vote::{self, UserVote, VoteState},
   },
-  DbError,
+  queries, DbError,
 };
 use futures::{select, FutureExt};
 use tokio::spawn;
@@ -44,10 +44,11 @@ pub async fn create_comment(
   assert_authenticated(&auth_session)?;
   // todo: item is dead
   // assert item exists?
-  let item =
-    db::get_item_by_id(&state.pool, payload.parent_item_id).await?.ok_or(RouteError::NotFound)?;
+  let item = queries::get_item_by_id(&state.pool, payload.parent_item_id)
+    .await?
+    .ok_or(RouteError::NotFound)?;
   let new_comment: Comment = payload.try_into()?;
-  db::insert_comment(&state.pool, &new_comment).await?;
+  queries::insert_comment(&state.pool, &new_comment).await?;
 
   Ok(StatusCode::CREATED)
 }
@@ -64,12 +65,12 @@ pub async fn get_comment(
   auth_session: AuthSession,
 ) -> ApiResult<(Json<Comment>, Json<Option<VoteState>>)> {
   let comment =
-    db::get_comment_by_id(&state.pool, comment_id).await?.ok_or(RouteError::NotFound)?;
+    queries::get_comment_by_id(&state.pool, comment_id).await?.ok_or(RouteError::NotFound)?;
 
   match auth_session.user {
     Some(user) => {
       let user_name = &user.0.username;
-      let user_vote = db::get_user_vote_by_content_id(&state.pool, user_name, comment_id)
+      let user_vote = queries::get_user_vote_by_content_id(&state.pool, user_name, comment_id)
         .await
         .context("no vote found")?;
       let vote_state = user_vote.map(|v| v.vote_state);
@@ -104,10 +105,10 @@ pub async fn update_comment_vote(
   let user_name = &auth_session.user.unwrap().0.username;
 
   let (comment, user_vote) = {
-    let comment_task = db::get_comment_by_id(&state.pool, comment_id);
-    let user_vote_task = db::get_user_vote_by_content_id(&state.pool, user_name, comment_id);
+    let comment_task = queries::get_comment_by_id(&state.pool, comment_id);
+    let user_vote_task = queries::get_user_vote_by_content_id(&state.pool, user_name, comment_id);
     let (comment_result, maybe_user_vote) = tokio::try_join!(comment_task, user_vote_task)?;
-    let comment = comment_result.context("failed to query db for comment")?;
+    let comment = comment_result.context("failed to query queries for comment")?;
     (comment, maybe_user_vote)
   };
 
@@ -120,7 +121,7 @@ pub async fn update_comment_vote(
   }
 
   // create a new UserVote and increment the comment author's karma
-  db::submit_comment_vote(&mut state.pool, comment_id, user_name, parent_item_id, vote_state)
+  queries::submit_comment_vote(&mut state.pool, comment_id, user_name, parent_item_id, vote_state)
     .await?;
 
   Ok(StatusCode::OK)
@@ -136,11 +137,11 @@ pub async fn update_comment_favorite(
   let user_name = &auth_session.user.unwrap().0.username;
 
   let (comment, maybe_favorite) = {
-    let comment_task = db::get_comment_by_id(&state.pool, comment_id);
+    let comment_task = queries::get_comment_by_id(&state.pool, comment_id);
     let favorite_task =
-      db::get_user_favorite_by_username_and_item_id(&state.pool, user_name, comment_id);
+      queries::get_user_favorite_by_username_and_item_id(&state.pool, user_name, comment_id);
     let (comment_result, maybe_favorite) = tokio::try_join!(comment_task, favorite_task)?;
-    let comment = comment_result.context("failed to query db for comment")?;
+    let comment = comment_result.context("failed to query queries for comment")?;
     (comment, maybe_favorite)
   };
 
@@ -157,7 +158,7 @@ pub async fn update_comment_favorite(
   }
 
   // update favorite
-  db::insert_or_delete_user_favorite_for_comment(
+  queries::insert_or_delete_user_favorite_for_comment(
     &state.pool,
     user_name,
     maybe_favorite,
