+++ db/src/queries/mod.rs
@@ -1,0 +1,7 @@
+pub mod comments;
+pub mod items;
+pub mod user_favorites;
+pub mod user_votes;
+pub mod users;
+
+pub use self::{comments::*, items::*, user_favorites::*, user_votes::*, users::*};
