+++ db/src/queries/user_favorites.rs
